requirements.txt
django
flake8
djangorestframework
gunicorn
mysqlclient